+++ website/src/sdk/server_pb.ts
@@ -110,6 +110,8 @@ export declare namespace InfoRes {
 		hostVpnIp: string,
 		metadataEnabled: boolean,
 		isAdmin: boolean,
+		allowedIps: string,
+		dnsEnabled: boolean,
 	}
 }
 
@@ -173,6 +175,22 @@ export class InfoRes extends jspb.Message {
 		(jspb.Message as any).setProto3BooleanField(this, 6, value);
 	}
 
+	getAllowedIps(): string {
+		return jspb.Message.getFieldWithDefault(this, 7, "");
+	}
+
+	setAllowedIps(value: string): void {
+		(jspb.Message as any).setProto3StringField(this, 7, value);
+	}
+
+	getDnsEnabled(): boolean {
+		return jspb.Message.getFieldWithDefault(this, 8, false);
+	}
+
+	setDnsEnabled(value: boolean): void {
+		(jspb.Message as any).setProto3BooleanField(this, 8, value);
+	}
+
 	serializeBinary(): Uint8Array {
 		const writer = new jspb.BinaryWriter();
 		InfoRes.serializeBinaryToWriter(this, writer);
@@ -187,6 +205,8 @@ export class InfoRes extends jspb.Message {
 			hostVpnIp: this.getHostVpnIp(),
 			metadataEnabled: this.getMetadataEnabled(),
 			isAdmin: this.getIsAdmin(),
+			allowedIps: this.getAllowedIps(),
+			dnsEnabled: this.getDnsEnabled(),
 			
 		};
 	}
@@ -215,6 +235,14 @@ export class InfoRes extends jspb.Message {
 		const field6 = message.getIsAdmin();
 		if (field6 != false) {
 			writer.writeBool(6, field6);
+		}
+		const field7 = message.getAllowedIps();
+		if (field7.length > 0) {
+			writer.writeString(7, field7);
+		}
+		const field8 = message.getDnsEnabled();
+		if (field8 != false) {
+			writer.writeBool(8, field8);
 		}
 	}
 
@@ -256,6 +284,14 @@ export class InfoRes extends jspb.Message {
 				const field6 = reader.readBool()
 				message.setIsAdmin(field6);
 				break;
+			case 7:
+				const field7 = reader.readString()
+				message.setAllowedIps(field7);
+				break;
+			case 8:
+				const field8 = reader.readBool()
+				message.setDnsEnabled(field8);
+				break;
 			default:
 				reader.skipField();
 				break;
@@ -286,6 +322,8 @@ function InfoResFromObject(obj: InfoRes.AsObject | undefined): InfoRes | undefin
 	message.setHostVpnIp(obj.hostVpnIp);
 	message.setMetadataEnabled(obj.metadataEnabled);
 	message.setIsAdmin(obj.isAdmin);
+	message.setAllowedIps(obj.allowedIps);
+	message.setDnsEnabled(obj.dnsEnabled);
 	return message;
 }